Refactor Redis tests.
Check Redis directly to see whether the key has expired.  This should
avoid adding fudge factors to the test duration, although fudge is
still necessary to compensate for Redis' 1-second expiration
resolution.

Also expire the proper key out of the store.  When the middleware is
being used, the route key has the remotekey added to it.  Since that's
not strictly known until a request is made, I've put in the localhost
IP address as that's what express is giving me for a remote address on
my box.  No clue if that holds up universally.

Also tried to make the logic a little cleaner, although it's still far
from perfect.