Mochatize the memory tests.
I'm using a single handler for all these tests. This introduces
potential for one test to affect other tests (i.e. it's not very
"unit-y"). I think the gain in showing resiliancy during multiple uses
is worth the downside.

Removed the timeout from the Redis test as the interval is now a
second and it really should be able to end in under 2.